Blog-9: Chors -> cleanup blogs service functions

+++ src/blogs/service/blogs.service.ts
@@ -1,11 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { from, Observable, pipe } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
-import { Repository } from 'typeorm';
-import { CreateBlogDto } from '../model/blog.dto';
+import { DeleteResult, Repository } from 'typeorm';
+import { CreateBlogDto, UpdateBlogDto } from '../model/blog.dto';
 import { BlogEntity } from '../model/blog.entity';
-import { BlogEntry } from '../model/blog.interface';
+import { BlogEntryInterface } from '../model/blog.interface';
 
 @Injectable()
 export class BlogService {
@@ -14,30 +14,34 @@ export class BlogService {
     private readonly blogRepository: Repository<BlogEntity>,
   ) {}
 
-  async findAll(): Promise<Observable<BlogEntry[]>> {
+  async findAll(): Promise<Observable<BlogEntryInterface[]>> {
     return await from(this.blogRepository.find({ relations: ['author_id'] }));
   }
 
-  async findOne(id: string): Promise<Observable<BlogEntry>> {
-    return await from(
+  async findOne(id: string): Promise<Observable<BlogEntryInterface>> {
+    const blog = await from(
       this.blogRepository.findOne({ id }, { relations: ['author_id'] }),
     );
+    if (!blog) throw new NotFoundException();
+    return blog;
   }
 
-  async create(blogEntry: CreateBlogDto): Promise<Observable<BlogEntry>> {
+  async create(blogEntry: CreateBlogDto): Promise<Observable<BlogEntryInterface>> {
     return await from(this.blogRepository.save(blogEntry));
   }
 
   async updateOne(
     id: string,
-    blogEntry: BlogEntry,
-  ): Promise<Observable<Observable<BlogEntry>>> {
-    return await from(this.blogRepository.update(id, blogEntry)).pipe(
+    blogEntry: Partial<UpdateBlogDto>,
+  ): Promise<Observable<Observable<Partial<BlogEntryInterface>>>> {
+    return from(this.blogRepository.update(id, blogEntry)).pipe(
       switchMap(() => this.findOne(id)),
     );
   }
 
-  async deleteOne(id: string): Promise<Observable<any>> {
+  async deleteOne(id: string): Promise<Observable<DeleteResult>> {
+    const blog = await this.blogRepository.findOne(id)
+    if (!blog) throw new NotFoundException()
     return await from(this.blogRepository.delete(id));
   }
 }
